# swaggerLog.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ecommerce API",
    "version": "1.0.0",
    "description": "The Demo ecommerce api made for user, cart, and order transactions for an art products store."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Main server"
    }
  ],
  "paths": {
...
    "/api/products/{productId}": {
      "get": {
        "summary": "Get the product object by product id.",
        "description": "This path is using id parameter. Product object is fetched from database.",
        "operationId": "get_product_id",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "The id of the product.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns product with id matched to the id parameter.",
            "content": {
              "application/json": {
                "example": {
                  "id": 2,
                  "name": "Poires et Raisins",
                  "author": "Pascal Giroud",
                  "description": "Nature is my greatest source of inspiration.",
                  "category": "Impressionistic",
                  "image_url": "https://d3rf6j5nx5r04a.cloudfront.net/m-1L9HX9Ok69cgTlozcISIJ6efo=/1120x0/product/c/5/f0d3668ec9dc46c3b9f0078195522fcd_opt.jpg",
                  "status": "Active",
                  "price": 27021
                }
              }
            }
          },
          "400": {
            "description": "Error for inavlid product id.",
            "content": {
              "text/plain": {
                "examples": {
                  "Error": {
                    "value": "Invalid product number"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/categories/{category}": {
      "get": {
        "summary": "Get products in specified category.",
        "description": "This path fetches all products in a category using the category parameter.",
        "operationId": "get_products_categories",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "description": "Category of products",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all product objects in the selected category, to an array.",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 14,
                    "name": "Kinko Sennin on a Koi",
                    "author": "Netsuke",
                    "description": "Kinko Sennin on a Koi. Ivory, Meiji-Period (1868 - 1912)",
                    "category": "Netsuke",
                    "image_url": "https://woodblock-print.eu/images/stories/virtuemart/product/CF_N67_Unsterblicher%20mit%20Koi7.jpg",
                    "status": "Sold",
                    "price": 13640
                  },
                  {
                    "id": 15,
                    "name": "Dragon",
                    "author": "Netsuke",
                    "description": "Dragon. Ivory, Meiji-Period (1868 - 1912)",
                    "category": "Netsuke",
                    "image_url": "https://springfieldmuseums.org/wp-content/uploads/2021/01/GWVS_75_23_127_3_sm.jpg",
                    "status": "Active",
                    "price": 76499
                  },
                  {
                    "id": 16,
                    "name": "Raiden, the god of wind, thunder and lightning",
                    "author": "Netsuke",
                    "description": "Netsuke depicting Raiden, the god of wind, thunder and lightning, Japan, 19th century, carved ivory",
                    "category": "Netsuke",
                    "image_url": "http://www.vam.ac.uk/__data/assets/image/0017/236222/2006AL0460_netsuke_main.jpg",
                    "status": "Sold",
                    "price": 10634
                  },
                  "..."
                ]
              }
            }
          },
          "400": {
            "description": "Error for inavlid product category.",
            "content": {
              "text/plain": {
                "examples": {
                  "Error": {
                    "value": "Invalid category"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}